refactor: put default logging messages in Logger
