+++ src/core/types/arbitrageloops/loops/liqMempoolLoop.ts
@@ -1,13 +1,13 @@
 import { fromUtf8 } from "@cosmjs/encoding";
 import { StdFee } from "@cosmjs/stargate";
 
-import { getliqudationMessage } from "../../../chains/defaults/messages/getLiquidationMessage";
-import { initLiquidationOverseers } from "../../../chains/defaults/queries/initOverseers";
-import { tryLiquidationArb } from "../../arbitrage/arbitrage";
-import { ChainOperator } from "../../chainOperator/chainoperator";
-import { Logger } from "../../logging";
-import { LiquidationConfig } from "../base/configs";
-import { decodeMempool, IgnoredAddresses, Mempool, MempoolTx } from "../base/mempool";
+import { getliqudationMessage } from "../../../../chains/defaults/messages/getLiquidationMessage";
+import { initLiquidationOverseers } from "../../../../chains/defaults/queries/initOverseers";
+import { tryLiquidationArb } from "../../../arbitrage/arbitrage";
+import { ChainOperator } from "../../../chainOperator/chainoperator";
+import { Logger } from "../../../logging";
+import { LiquidationConfig } from "../../base/configs";
+import { decodeMempool, IgnoredAddresses, Mempool, MempoolTx } from "../../base/mempool";
 import {
 	adjustCollateral,
 	AnchorOverseer,
@@ -15,14 +15,14 @@ import {
 	repayStable,
 	setBorrowLimits,
 	setPriceFeed,
-} from "../base/overseer";
+} from "../../base/overseer";
 import {
 	isBorrowStableMessage,
 	isLockCollateralMessage,
 	isRepayStableMessage,
 	isUnlockCollateralMessage,
 	PriceFeedMessage,
-} from "../messages/liquidationmessages";
+} from "../../messages/liquidationmessages";
 /*
  *
  */
